Fix migration breaking local fresh migrate

+++ cncnet-api/database/migrations/2024_04_15_100832_add_d2k_countable_game_objects.php
@@ -15,6 +15,10 @@ return new class extends Migration
     {
         //
         $gameObjectSchema = GameObjectSchema::find(4); // d2k
+        if ($gameObjectSchema == null)
+        {
+            return;
+        }
         $gameObjects = CountableGameObject::where("game_object_schema_id", 4)->delete();
 
         $d2k_units = [
